# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARIA GOLD AI V3 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  agentId: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export interface ConversationInput {
  title: string;
  agentId?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  tokensUsed?: number | null;
  createdAt: string;
}

export interface MessageInput {
  content: string;
  agentId?: string;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export interface Agent {
  id: string;
  name: string;
  role: string;
  description: string;
  status: string;
  goals: string[];
  tools: string[];
  /** @nullable */
  tasksCompleted?: number | null;
  /** @nullable */
  lastActive?: string | null;
}

export interface Memory {
  id: number;
  type: string;
  content: string;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  conversationId?: number | null;
  /** @nullable */
  repositoryId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface MemoryInput {
  type: string;
  content: string;
  key?: string;
  conversationId?: number;
  repositoryId?: number;
}

export interface MemoryUpdate {
  content?: string;
  key?: string;
}

export interface Repository {
  id: number;
  name: string;
  url: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  branch?: string | null;
  status?: string;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  securityScore?: number | null;
  /** @nullable */
  performanceScore?: number | null;
  /** @nullable */
  maintainabilityScore?: number | null;
  /** @nullable */
  issuesFound?: number | null;
  /** @nullable */
  lastAnalyzedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RepositoryInput {
  name: string;
  url: string;
  description?: string;
  branch?: string;
}

export interface AgentHealthItem {
  agentId: string;
  name: string;
  status: string;
  tasksCompleted: number;
}

export interface OwnerStats {
  systemHealth: string;
  totalTokensUsed: number;
  totalCostUsd: number;
  activeAgents: number;
  totalConversations: number;
  totalRepositories: number;
  agentHealth: AgentHealthItem[];
}

export interface UsageDataPoint {
  date: string;
  tokensUsed: number;
  costUsd: number;
  requestCount: number;
}

export type ListMemoryParams = {
query?: string;
type?: string;
};