# lib/api-zod/src/generated/types/memory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARIA GOLD AI V3 API
 * OpenAPI spec version: 0.1.0
 */

export interface Memory {
  id: number;
  type: string;
  content: string;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  conversationId?: number | null;
  /** @nullable */
  repositoryId?: number | null;
  createdAt: string;
  updatedAt: string;
}