# lib/api-zod/src/generated/types/memoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARIA GOLD AI V3 API
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryInput {
  type: string;
  content: string;
  key?: string;
  conversationId?: number;
  repositoryId?: number;
}